requirements.txt
PyQt6>=6.6.0
paddlepaddle==3.0.0
paddleocr>=2.7.0.3
pymupdf>=1.23.0
Pillow>=10.0.0
numpy>=1.24.0
rapidfuzz>=3.0.0
opencv-python-headless>=4.8.0
pyinstaller>=6.0.0
pyyaml>=6.0
openpyxl>=3.1.0
